[bug] range i32

+++ include/pathfinding.h
@@ -149,20 +149,20 @@ void Pathfinding_Visible_noM(i32 *sightmap, i32 *blockmap, size_t row_len,
 // - assailable
 // TODO: attack to use position and range instead of movemap
 i32 *Pathfinding_Attackto(i32 *move_matrix, size_t row_len, size_t col_len,
-                          u8 range[2], i32 mode_movetile);
+                          i32 range[2], i32 mode_movetile);
 void Pathfinding_Attackto_noM(i32 *attackto_mat, i32 *move_matrix,
                               size_t row_len, size_t col_len,
-                              u8 range[2], i32 mode_movetile);
+                              i32 range[2], i32 mode_movetile);
 void _Pathfinding_Attackto(i32 x, i32 y, i32 *attackmap, i32 *movemat,
                            size_t row_len, size_t col_len,
-                           u8 range[2], i32 mode_movetile);
+                           i32 range[2], i32 mode_movetile);
 
 /* -- Attackfrom -- */
 i32 *Pathfinding_Attackfrom(i32 *movemap, size_t row_len, size_t col_len,
-                            struct Point target, u8 range[2], int mode_output);
+                            struct Point target, i32 range[2], int mode_output);
 void Pathfinding_Attackfrom_noM(i32 *attackfrom, i32 *movemap,
                                 size_t row_len, size_t col_len,
-                                struct Point target, u8 range[2]);
+                                struct Point target, i32 range[2]);
 
 /* -- Utils -- */
 typedef i32 *(*PathList_f)(i32 *, i32 *, size_t, struct Point, struct Point );
